# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Singers catalog API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/singers": {
            "get": {
                "tags": [
                    "Singers"
                ],
                "summary": "Singers list",
                "operationId": "3c32ee2b0e2575e54a8519dce348eeb6",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Singer name"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-18T16:58:51.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-18T17:09:39.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Singers"
                ],
                "summary": "Create singer",
                "operationId": "6866b1eb30d7995b22205def483a1863",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Singer name"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Singer name"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T16:58:51.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T17:09:39.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/singers/{singer}": {
            "delete": {
                "tags": [
                    "Singers"
                ],
                "summary": "Delete singer",
                "operationId": "d3f829a2339b31dc89cec0760af16ed6",
                "parameters": [
                    {
                        "name": "singer",
                        "in": "path",
                        "description": "Singer ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": "Successfully deleted"
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Singers"
                ],
                "summary": "Update singer",
                "operationId": "7265ca52f10776f2f19628fa378004ee",
                "parameters": [
                    {
                        "name": "singers",
                        "in": "path",
                        "description": "Song ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Singers name"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Singers name"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T16:58:51.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T17:09:39.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Song list",
                "operationId": "fb1c6430e8ade2f564b123895208d5dc",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Song name"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-18T16:58:51.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-18T17:09:39.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Create song",
                "operationId": "86b720816867b5a390573276df2a091d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Song name"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Song name"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T16:58:51.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T17:09:39.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/songs/{song}": {
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Delete song",
                "operationId": "07ac5e3b0305be25ea34a3a12af94033",
                "parameters": [
                    {
                        "name": "song",
                        "in": "path",
                        "description": "Song ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": "Successfully deleted"
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Songs"
                ],
                "summary": "Update song",
                "operationId": "c3b4b9800de9639aa7238bab812aa0d7",
                "parameters": [
                    {
                        "name": "song",
                        "in": "path",
                        "description": "Song ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Song name"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Song name"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T16:58:51.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T17:09:39.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}